move all the logic to a custom hook
